Test added, refactoring
