+++ activiti-cloud-starter-runtime-bundle/src/test/java/org/activiti/cloud/starter/tests/services/audit/AuditProducerIT.java
@@ -5,7 +5,6 @@ import static org.activiti.api.model.shared.event.VariableEvent.VariableEvents.V
 import static org.activiti.api.process.model.events.BPMNActivityEvent.ActivityEvents.ACTIVITY_CANCELLED;
 import static org.activiti.api.process.model.events.BPMNActivityEvent.ActivityEvents.ACTIVITY_COMPLETED;
 import static org.activiti.api.process.model.events.BPMNActivityEvent.ActivityEvents.ACTIVITY_STARTED;
-import static org.activiti.api.process.model.events.BPMNSignalEvent.SignalEvents.SIGNAL_RECEIVED;
 import static org.activiti.api.process.model.events.ProcessRuntimeEvent.ProcessEvents.PROCESS_CANCELLED;
 import static org.activiti.api.process.model.events.ProcessRuntimeEvent.ProcessEvents.PROCESS_COMPLETED;
 import static org.activiti.api.process.model.events.ProcessRuntimeEvent.ProcessEvents.PROCESS_CREATED;
@@ -42,7 +41,6 @@ import org.activiti.api.process.model.builders.ProcessPayloadBuilder;
 import org.activiti.api.process.model.builders.StartProcessPayloadBuilder;
 import org.activiti.api.process.model.events.BPMNActivityEvent;
 import org.activiti.api.process.model.events.ProcessDefinitionEvent;
-import org.activiti.api.process.model.payloads.SignalPayload;
 import org.activiti.api.task.model.Task;
 import org.activiti.api.task.model.TaskCandidateGroup;
 import org.activiti.api.task.model.TaskCandidateUser;
@@ -50,14 +48,11 @@ import org.activiti.api.task.model.builders.TaskPayloadBuilder;
 import org.activiti.cloud.api.model.shared.events.CloudRuntimeEvent;
 import org.activiti.cloud.api.process.model.CloudProcessDefinition;
 import org.activiti.cloud.api.process.model.CloudProcessInstance;
-import org.activiti.cloud.api.process.model.events.CloudBPMNActivityCompletedEvent;
 import org.activiti.cloud.api.process.model.events.CloudBPMNActivityStartedEvent;
 import org.activiti.cloud.api.process.model.events.CloudProcessDeployedEvent;
 import org.activiti.cloud.api.task.model.CloudTask;
-import org.activiti.cloud.api.task.model.events.CloudTaskAssignedEvent;
 import org.activiti.cloud.api.task.model.events.CloudTaskCancelledEvent;
 import org.activiti.cloud.api.task.model.events.CloudTaskCandidateUserRemovedEvent;
-import org.activiti.cloud.api.task.model.events.CloudTaskCreatedEvent;
 import org.activiti.cloud.starter.tests.helper.ProcessInstanceRestTemplate;
 import org.activiti.cloud.starter.tests.helper.SignalRestTemplate;
 import org.activiti.cloud.starter.tests.helper.TaskRestTemplate;
@@ -89,12 +84,7 @@ public class AuditProducerIT {
     private static final String SIMPLE_SUB_PROCESS1 = "simpleSubProcess1";
     private static final String SIMPLE_SUB_PROCESS2 = "simpleSubProcess2";
     private static final String CALL_TWO_SUB_PROCESSES = "callTwoSubProcesses";
-    private static final String SIMPLE_EMBEDDED_SUB_PROCESS = "startSimpleSubProcess";
-    private static final String SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY = "startSimpleSubProcessWithCallActivity";
-    private static final String SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT = "signalSubProcess";
-    
-    
-    
+       
     public static final String ROUTING_KEY_HEADER = "routingKey";
     public static final String[] RUNTIME_BUNDLE_INFO_HEADERS = {"appName", "appVersion", "serviceName", "serviceVersion", "serviceFullName", ROUTING_KEY_HEADER};
     public static final String[] ALL_REQUIRED_HEADERS = Stream.of(RUNTIME_BUNDLE_INFO_HEADERS)
@@ -114,10 +104,6 @@ public class AuditProducerIT {
     @Autowired
     private ProcessInstanceRestTemplate processInstanceRestTemplate;
     
-    @Autowired
-    private SignalRestTemplate signalRestTemplate;
-    
-
     @Autowired
     private TaskRestTemplate taskRestTemplate;
 
@@ -657,316 +643,7 @@ public class AuditProducerIT {
         runtimeService.deleteProcessInstance(processInstanceId, "Clean up");
 
     }
-
-    @Test
-    public void testEmbeddedSubProcesses() {
-        //given
-        ResponseEntity<CloudProcessInstance> processInstance = processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_EMBEDDED_SUB_PROCESS));
-
-        String processInstanceId = processInstance.getBody().getId();
-        
-        await().untilAsserted(() -> {
-            
-          assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-            
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getLatestReceivedEvents();
-            
-
-            assertThat(receivedEvents)
-                    .extracting(CloudRuntimeEvent::getEventType,
-                                CloudRuntimeEvent::getProcessInstanceId,
-                                CloudRuntimeEvent::getParentProcessInstanceId,
-                                CloudRuntimeEvent::getProcessDefinitionKey)
-                    .containsExactly(tuple(PROCESS_CREATED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(PROCESS_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(ACTIVITY_COMPLETED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(SEQUENCE_FLOW_TAKEN,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(ACTIVITY_COMPLETED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(SEQUENCE_FLOW_TAKEN,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(TASK_CANDIDATE_GROUP_ADDED,null, null, null),
-                                     tuple(TASK_CREATED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS)
-                    );
-            
-            assertThat(receivedEvents)
-                .filteredOn(event -> TASK_CREATED.equals(event.getEventType()))
-                .extracting(CloudRuntimeEvent::getProcessInstanceId,
-                            event -> ((CloudTaskCreatedEvent) event).getEntity().getProcessInstanceId())
-                .containsExactly(tuple(processInstanceId, processInstanceId));
-            
-            
-           List<CloudRuntimeEvent<?, ?>> activitiStartedEvents = receivedEvents.stream()
-                                                                    .filter(event -> ACTIVITY_STARTED.equals(event.getEventType()))
-                                                                    .collect(Collectors.toList());
-            
-           assertThat(activitiStartedEvents)
-               .filteredOn(event -> ACTIVITY_STARTED.equals(event.getEventType()))
-               .extracting(event -> ((CloudBPMNActivityStartedEvent) event).getEntity().getActivityType(),
-                           event -> ((CloudBPMNActivityStartedEvent) event).getEntity().getActivityName())
-               .containsExactly(tuple("startEvent", null),
-                                tuple("subProcess", "subProcess"),
-                                tuple("startEvent", null),
-                                tuple("userTask", "Task in subprocess"));
-            
-
-
-        });
-       
-        ResponseEntity<PagedResources<CloudTask>> tasks = processInstanceRestTemplate.getTasks(processInstance);
-        Task task = tasks.getBody().iterator().next();
-        
-        //when
-        taskRestTemplate.claim(task);
-        
-        await().untilAsserted(() -> {
-        
-            assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getLatestReceivedEvents();
-            
-            assertThat(receivedEvents)
-                    .extracting(CloudRuntimeEvent::getEventType,
-                                CloudRuntimeEvent::getProcessInstanceId,
-                                CloudRuntimeEvent::getParentProcessInstanceId,
-                                CloudRuntimeEvent::getProcessDefinitionKey
-                                )
-                    .containsExactly(tuple(TASK_ASSIGNED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS),
-                                     tuple(TASK_UPDATED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS));
-
-            
-            String entityProcessInstanceId = ((CloudTaskAssignedEvent) receivedEvents.get(0)).getEntity().getProcessInstanceId();
-            assertThat(entityProcessInstanceId).isNotNull();
-            assertThat(entityProcessInstanceId).isEqualTo(processInstanceId);
-            
-        });
-        
-        //when
-        taskRestTemplate.complete(task);
       
-        //then
-        await().untilAsserted(() -> {
-          assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-          List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getLatestReceivedEvents();
-
-          assertThat(receivedEvents)
-                  .extracting(CloudRuntimeEvent::getEventType,
-                      CloudRuntimeEvent::getProcessInstanceId,
-                      CloudRuntimeEvent::getProcessDefinitionKey)
-                  .containsExactly(tuple(TASK_COMPLETED,processInstanceId,SIMPLE_EMBEDDED_SUB_PROCESS),
-                                   tuple(TASK_CANDIDATE_GROUP_REMOVED,null,null),
-                                   tuple(ACTIVITY_COMPLETED,processInstanceId,SIMPLE_EMBEDDED_SUB_PROCESS),
-                                   tuple(SEQUENCE_FLOW_TAKEN,processInstanceId,SIMPLE_EMBEDDED_SUB_PROCESS),
-                                   tuple(ACTIVITY_STARTED,processInstanceId,SIMPLE_EMBEDDED_SUB_PROCESS),
-                                   tuple(ACTIVITY_COMPLETED,processInstanceId,SIMPLE_EMBEDDED_SUB_PROCESS),
-                                   tuple(ACTIVITY_COMPLETED,processInstanceId,SIMPLE_EMBEDDED_SUB_PROCESS)/*subProcess*/,
-                                   tuple(SEQUENCE_FLOW_TAKEN,processInstanceId,SIMPLE_EMBEDDED_SUB_PROCESS),
-                                   tuple(ACTIVITY_STARTED,processInstanceId,SIMPLE_EMBEDDED_SUB_PROCESS),
-                                   tuple(ACTIVITY_COMPLETED,processInstanceId,SIMPLE_EMBEDDED_SUB_PROCESS),
-                                   tuple(PROCESS_COMPLETED,processInstanceId,SIMPLE_EMBEDDED_SUB_PROCESS));
-          
-          CloudBPMNActivityCompletedEvent subprocessCompletedEvent = (CloudBPMNActivityCompletedEvent)receivedEvents.stream()
-                  .filter(event -> ACTIVITY_COMPLETED.equals(event.getEventType()) && 
-                                  "subProcess".equals(((CloudBPMNActivityCompletedEvent) event).getEntity().getActivityType()))
-                  .collect(Collectors.toList())
-                  .get(0);
-          
-                  assertThat(subprocessCompletedEvent).isNotNull();
-                  assertThat(subprocessCompletedEvent.getProcessInstanceId()).isEqualTo(processInstanceId);
-        });
-
-    }
-    
-    
-    @Test
-    public void testEmbeddedSubProcessWithCallActivity() {
-        //given
-        ResponseEntity<CloudProcessInstance> processInstance = processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY));
-
-        String processInstanceId = processInstance.getBody().getId();
-        
-        await().untilAsserted(() -> {
-            
-          assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-            
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getLatestReceivedEvents();
-            
-
-            assertThat(receivedEvents)
-                    .extracting(CloudRuntimeEvent::getEventType,
-                                CloudRuntimeEvent::getProcessInstanceId,
-                                CloudRuntimeEvent::getParentProcessInstanceId,
-                                CloudRuntimeEvent::getProcessDefinitionKey)
-                    .containsExactly(tuple(PROCESS_CREATED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(PROCESS_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(ACTIVITY_COMPLETED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(SEQUENCE_FLOW_TAKEN,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(ACTIVITY_COMPLETED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(SEQUENCE_FLOW_TAKEN,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(TASK_CANDIDATE_GROUP_ADDED,null, null, null),
-                                     tuple(TASK_CREATED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY)
-                    );
-            
-            
-            assertThat(receivedEvents)
-            .filteredOn(event -> ACTIVITY_STARTED.equals(event.getEventType()))
-            .extracting(event -> ((CloudBPMNActivityStartedEvent) event).getEntity().getActivityType(),
-                        event -> ((CloudBPMNActivityStartedEvent) event).getEntity().getActivityName()
-                        )
-            .containsExactly(tuple("startEvent", null),
-                             tuple("subProcess", "subProcess"),
-                             tuple("startEvent", null),
-                             tuple("userTask", "Task in subprocess")
-                             );
-            
-
-        });
-       
-        ResponseEntity<PagedResources<CloudTask>> tasks = processInstanceRestTemplate.getTasks(processInstance);
-        Task task = tasks.getBody().iterator().next();
-        
-        String subProcessInstanceId = task.getProcessInstanceId();
-        assertThat(subProcessInstanceId).isNotNull();
-        assertThat(subProcessInstanceId).isEqualTo(processInstanceId);
-        
-        //when
-        taskRestTemplate.claim(task);
-        
-        await().untilAsserted(() -> {
-            
-            assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getLatestReceivedEvents();
-            
-            assertThat(receivedEvents)
-                    .extracting(CloudRuntimeEvent::getEventType,
-                                CloudRuntimeEvent::getProcessInstanceId,
-                                CloudRuntimeEvent::getParentProcessInstanceId,
-                                CloudRuntimeEvent::getProcessDefinitionKey
-                                )
-                    .containsExactly(tuple(TASK_ASSIGNED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(TASK_UPDATED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY));
-        });
-
-        
-        //when
-        taskRestTemplate.complete(task);
-        
-        //Check we have two processes and one of them is callActivity process
-        // when
-        List<org.activiti.engine.runtime.ProcessInstance> processes = runtimeService.createProcessInstanceQuery()
-                                                   .superProcessInstanceId(processInstanceId)
-                                                   .list()
-                                                   .stream()
-                                                   .collect(Collectors.toList());
-        assertThat(processes).hasSize(1);
-        String callActivityProcessId = processes.get(0).getProcessInstanceId();
-        
-        assertThat(processes.get(0).getProcessDefinitionKey()).isEqualTo(SIMPLE_SUB_PROCESS1);
-  
-        await().untilAsserted(() -> {
-        
-            assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getLatestReceivedEvents();
-           
-            assertThat(receivedEvents)
-                    .extracting(CloudRuntimeEvent::getEventType,
-                                CloudRuntimeEvent::getProcessInstanceId,
-                                CloudRuntimeEvent::getParentProcessInstanceId,
-                                CloudRuntimeEvent::getProcessDefinitionKey
-                                )
-                    .containsExactly(tuple(TASK_COMPLETED,processInstanceId,null,SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(TASK_CANDIDATE_GROUP_REMOVED,null,null,null),
-                                     tuple(ACTIVITY_COMPLETED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(SEQUENCE_FLOW_TAKEN, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(ACTIVITY_STARTED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_CALLACTIVITY),
-                                     tuple(PROCESS_CREATED, callActivityProcessId, processInstanceId, SIMPLE_SUB_PROCESS1),
-                                     tuple(PROCESS_STARTED, callActivityProcessId, processInstanceId, SIMPLE_SUB_PROCESS1),
-                                     tuple(ACTIVITY_STARTED, callActivityProcessId, processInstanceId, SIMPLE_SUB_PROCESS1),
-                                     tuple(ACTIVITY_COMPLETED, callActivityProcessId, processInstanceId, SIMPLE_SUB_PROCESS1),
-                                     tuple(SEQUENCE_FLOW_TAKEN, callActivityProcessId, processInstanceId, SIMPLE_SUB_PROCESS1),
-                                     tuple(ACTIVITY_STARTED, callActivityProcessId, processInstanceId, SIMPLE_SUB_PROCESS1),
-                                     tuple(TASK_CREATED, callActivityProcessId, processInstanceId, SIMPLE_SUB_PROCESS1)
-                            );
-
-            
-        });
-        
-
-        // Clean up
-        runtimeService.deleteProcessInstance(callActivityProcessId, "Clean up");
-        runtimeService.deleteProcessInstance(processInstanceId, "Clean up");
- 
-    }
-      
-    @Test
-    public void testEmbeddedSubProcessWithSignalEvent() {
-        //given
-        ResponseEntity<CloudProcessInstance> processInstance = processInstanceRestTemplate.startProcess(processDefinitionIds.get(SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT));
-
-        String processInstanceId = processInstance.getBody().getId();
-        
-        await().untilAsserted(() -> {
-            
-            assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getLatestReceivedEvents();
-            assertThat(receivedEvents)
-                    .extracting(CloudRuntimeEvent::getEventType,
-                                CloudRuntimeEvent::getProcessInstanceId,
-                                CloudRuntimeEvent::getParentProcessInstanceId,
-                                CloudRuntimeEvent::getProcessDefinitionKey)
-                    .containsExactly(tuple(PROCESS_CREATED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(PROCESS_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(ACTIVITY_COMPLETED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(SEQUENCE_FLOW_TAKEN, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(SEQUENCE_FLOW_TAKEN, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(TASK_CREATED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(ACTIVITY_COMPLETED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(SEQUENCE_FLOW_TAKEN, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                                     tuple(ACTIVITY_STARTED,processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT)
-                    );
-        });
-    
-        SignalPayload signalProcessInstancesCmd = ProcessPayloadBuilder
-              .signal()
-              .withName("mySignal")
-              .build();
-    
-        //when
-        signalRestTemplate.signal(signalProcessInstancesCmd);   
-
-        await().untilAsserted(() -> {
-            
-            assertThat(streamHandler.getReceivedHeaders()).containsKeys(ALL_REQUIRED_HEADERS);
-            List<CloudRuntimeEvent<?, ?>> receivedEvents = streamHandler.getLatestReceivedEvents();
-              
-            assertThat(receivedEvents)
-            .extracting(CloudRuntimeEvent::getEventType,
-                        CloudRuntimeEvent::getProcessInstanceId,
-                        CloudRuntimeEvent::getParentProcessInstanceId,
-                        CloudRuntimeEvent::getProcessDefinitionKey)
-            .containsExactly(tuple(SIGNAL_RECEIVED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                             tuple(ACTIVITY_COMPLETED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                             tuple(SEQUENCE_FLOW_TAKEN, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                             tuple(ACTIVITY_STARTED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                             tuple(ACTIVITY_COMPLETED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                             tuple(ACTIVITY_COMPLETED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                             tuple(SEQUENCE_FLOW_TAKEN, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                             tuple(ACTIVITY_STARTED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT),
-                             tuple(ACTIVITY_COMPLETED, processInstanceId, null, SIMPLE_EMBEDDED_SUB_PROCESS_WITH_SIGNAL_EVENT)
-            );
-        });
-        
-        runtimeService.deleteProcessInstance(processInstanceId, "Clean up");        
-    }
-
     private ResponseEntity<PagedResources<CloudProcessDefinition>> getProcessDefinitions() {
         ParameterizedTypeReference<PagedResources<CloudProcessDefinition>> responseType = new ParameterizedTypeReference<PagedResources<CloudProcessDefinition>>() {
         };
